requirements.txt
asgiref==3.5.1
async-generator==1.10
attrs==21.4.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
black==22.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
Django==4.0.4
django-rest-framework==0.1.0
djangorestframework==3.13.1
fake-useragent==0.1.11
h11==0.13.0
httmock==1.4.0
http-request-randomizer==1.3.2
idna==3.3
isort==5.10.1
mypy-extensions==0.4.3
outcome==1.1.0
pathspec==0.9.0
platformdirs==2.5.2
psutil==5.9.0
pycparser==2.21
pydub==0.25.1
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
requests==2.27.1
selenium==4.1.5
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
sqlparse==0.4.2
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.2.0
urllib3==1.26.9
webdriver-manager==3.5.4
wsproto==1.1.0
